Constants should be readonly.
Reading the log should accept a ongoing write lock, given that the client keeps this lock even if we halt logging in-game.

Split the Utils into sub-classes simply to spread and categorize them a bit better.

Some of the PathUtils can't throw anymore, but instead return null.

Console and Exit actions are all in the ProgramUtils to keep things gathered.

The namespace needed the ".src" (folder name) to avoid linter complaints.