Added a build section to the readme with a simple build command to build a 64-bit windows binary.
Reorganized the settings class and split it into env/arguments, then merged it together in the settings class.

Similar was done with the program code, some of it is now in the worker class which focuses on just reading and splitting logs.

Added fallback exception catch and log writing if something goes wrong and crashes, just in case.