Saved a lot of memory by simply making the LogReader also deal with the Gap handling.
The LogReader.IsClose evaluates if two timestamps are to be considered "the same log" or "split in a separate file".

This way we can also keep the List of timestamps limited to the amount of groups we need to later split into.

Previously it kept track of all the individual timestamps/lines, and later grouped and "filtered" which meant a lot of wasted RAM and CPU cycles, whops!

Simultaneously, fixed an issue with the Position tracking where the chars length would include new line bytes, which we don't want.

The position ought to be the start of the line, so the FileReader/FileWriter later can properly copy the range properly.

The cause is because the buffer has room for the largest possible timestamp, since the month and day can be either one or two digits, this meant sometimes the new line bytes got caught up in the count, which they shouldn't.

Got rid of several delegate syntax and simply using Action or Func when defining a callback.