When the log writer has split a group it then writes the info about it (instead of the full output before splitting starts).
The static ProgramInfo class will help provide Name and Version info for any platform by using the ProcessModule or the Assembly information, whichever contains information.

The IsTimestamp and ExtractTimestamp also have a secondary return which is offset information needed for proper position calculations.

The LogReader now has a AddLine method which deals with the adding or updating logic.

The OnByte method now properly calculates the position by applying the timestamp length, this way the position is always at the start of the line, even if there was new lines or other fluff in the buffer.